refactor: 이미지 크롭시 처음 포커스 되는 요소 조작을 통해 스크린리더가 올바르게 읽도록 변경
